requirements.txt
aiohttp==3.7.4.post0
alembic==1.6.5
anyio==3.2.1
apache-airflow==2.1.0
apache-airflow-providers-cncf-kubernetes==2.0.0
apache-airflow-providers-ftp==2.0.0
apache-airflow-providers-hashicorp==2.0.0
apache-airflow-providers-imap==2.0.0
apache-airflow-providers-postgres==2.0.0
apache-airflow-providers-slack==4.0.0
apache-airflow-providers-sqlite==2.0.0
apispec==3.3.2
argcomplete==1.12.3
asn1crypto==1.4.0
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.1
blinker==1.4
cachetools==4.2.2
cattrs==1.7.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==5.0.1
commonmark==0.9.1
croniter==1.0.15
cryptography==3.4.7
defusedxml==0.7.1
dill==0.3.4
dnspython==2.1.0
docutils==0.17.1
email-validator==1.1.3
Flask==1.1.4
Flask-AppBuilder==3.3.1
Flask-Babel==1.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
google-auth==1.32.0
graphviz==0.16
gunicorn==20.1.0
h11==0.12.0
helpers==0.2.0
httpcore==0.13.6
httpx==0.18.2
hvac==0.10.14
idna==3.2
importlib-resources==1.5.0
inflection==0.5.1
iso8601==0.1.14
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
kubernetes==11.0.0
lazy-object-proxy==1.6.0
lockfile==0.12.2
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.12.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==2.1.0
marshmallow-sqlalchemy==0.23.1
multidict==5.1.0
numpy==1.21.0
oauthlib==3.1.1
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
pandas==1.2.5
pendulum==2.1.2
pg8000==1.20.0
postgres==3.0.0
prison==0.1.3
psutil==5.8.0
psycopg2-binary==2.9.1
psycopg2-pool==1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.9.0
pyhcl==0.4.4
PyJWT==1.7.1
pyrsistent==0.17.3
pyslack==0.5.0
python-daemon==2.3.0
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rfc3986==1.5.0
rich==10.4.0
rsa==4.7.2
scramp==1.4.0
setproctitle==1.2.2
six==1.16.0
slack==0.0.2
slack-sdk==3.7.0
slackclient==2.9.3
sniffio==1.2.0
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.8
swagger-ui-bundle==0.0.8
tabulate==0.8.9
tenacity==6.2.0
termcolor==1.1.0
testing.common.database==2.0.3
text-unidecode==1.3
typing-extensions==3.10.0.0
unicodecsv==0.14.1
urllib3==1.26.6
websocket-client==1.1.0
Werkzeug==1.0.1
WTForms==2.3.3
yarl==1.6.3
